feat/#56: OutlinedSecondaryButton Component
