+++ core/designsystem/src/main/java/com/yapp/core/designsystem/component/button/outlined/OutlinedButton.kt
@@ -56,87 +56,3 @@ fun YappOutlinedButtonBasic(
         }
     }
 }
-
-@Composable
-fun YappOutlinedPrimaryButtonXLarge(
-    modifier: Modifier = Modifier,
-    shape: Shape = OutlinedButtonDefaults.shapeXLarge,
-    textStyle: TextStyle = OutlinedButtonDefaults.textStyleXLarge,
-    colors: OutlinedButtonColors = OutlinedButtonDefaults.colorsPrimary,
-    contentPaddings: PaddingValues = OutlinedButtonDefaults.contentPaddingsXLarge,
-    text: String,
-    enable: Boolean = true,
-    onClick: () -> Unit,
-) {
-    YappOutlinedButtonBasic(
-        modifier = modifier,
-        shape = shape,
-        text = text,
-        textStyle = textStyle,
-        colors = colors,
-        contentPaddings = contentPaddings,
-        enable = enable,
-        onClick = onClick
-    )
-}
-
-@Composable
-fun YappOutlinedPrimaryButtonLarge(
-    modifier: Modifier = Modifier,
-    shape: Shape = OutlinedButtonDefaults.shapeLarge,
-    textStyle: TextStyle = OutlinedButtonDefaults.textStyleLarge,
-    colors: OutlinedButtonColors = OutlinedButtonDefaults.colorsPrimary,
-    contentPaddings: PaddingValues = OutlinedButtonDefaults.contentPaddingsLarge,
-    text: String,
-    enable: Boolean = true,
-    onClick: () -> Unit,
-) {
-    YappOutlinedButtonBasic(
-        modifier = modifier,
-        shape = shape,
-        text = text,
-        textStyle = textStyle,
-        colors = colors,
-        contentPaddings = contentPaddings,
-        enable = enable,
-        onClick = onClick
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun YappOutlinedPrimaryButtonXLargePreview() {
-    YappTheme {
-        Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            YappOutlinedPrimaryButtonXLarge(
-                text = "Label",
-                enable = true,
-                onClick = {}
-            )
-            YappOutlinedPrimaryButtonXLarge(
-                text = "Label",
-                enable = false,
-                onClick = {}
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun YappOutlinedPrimaryButtonLargePreview() {
-    YappTheme {
-        Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            YappOutlinedPrimaryButtonLarge(
-                text = "Label",
-                enable = true,
-                onClick = {}
-            )
-            YappOutlinedPrimaryButtonLarge(
-                text = "Label",
-                enable = false,
-                onClick = {}
-            )
-        }
-    }
-}
